feat: Add Page component for dungeon editor
This commit adds a new component called Page to the dungeon editor feature. The Page component is responsible for rendering the main content of the editor, which includes the Editor component. The Editor component is imported from the playground folder.

The purpose of this change is to modularize the code and improve the organization of the dungeon editor feature.

Refactor: Remove unused use-modal hook
Refactor: Update setFloatingElemPosition to use _isLink parameter
Stupid "declared not use" lint rule!
Added initial serialization function